refactor(hook): 지도 관련 로직을 useMap 커스텀 훅으로 분리
